Move API code into orderbook crate
