+++ crates/orderbook/src/api/post_order.rs
@@ -1,19 +1,19 @@
 use {
-    crate::orderbook::{AddOrderError, Orderbook},
+    crate::{
+        api::{error, extract_payload, ApiReply, IntoWarpReply},
+        orderbook::{AddOrderError, Orderbook},
+    },
     anyhow::Result,
     model::{
         order::{AppdataFromMismatch, OrderCreation, OrderUid},
         quote::QuoteId,
         signature,
     },
-    shared::{
-        api::{error, extract_payload, ApiReply, IntoWarpReply},
-        order_validation::{
-            AppDataValidationError,
-            OrderValidToError,
-            PartialValidationError,
-            ValidationError,
-        },
+    shared::order_validation::{
+        AppDataValidationError,
+        OrderValidToError,
+        PartialValidationError,
+        ValidationError,
     },
     std::{convert::Infallible, sync::Arc},
     warp::{
@@ -88,7 +88,7 @@ impl IntoWarpReply for PartialValidationErrorWrapper {
             ),
             PartialValidationError::Other(err) => {
                 tracing::error!(?err, "PartialValidatonError");
-                shared::api::internal_error_reply()
+                crate::api::internal_error_reply()
             }
         }
     }
@@ -226,7 +226,7 @@ impl IntoWarpReply for ValidationErrorWrapper {
 
             ValidationError::Other(err) => {
                 tracing::error!(?err, "ValidationErrorWrapper");
-                shared::api::internal_error_reply()
+                crate::api::internal_error_reply()
             }
         }
     }
@@ -242,7 +242,7 @@ impl IntoWarpReply for AddOrderError {
             ),
             Self::Database(err) => {
                 tracing::error!(?err, "AddOrderError");
-                shared::api::internal_error_reply()
+                crate::api::internal_error_reply()
             }
             err @ AddOrderError::AppDataMismatch { .. } => {
                 tracing::error!(
@@ -251,7 +251,7 @@ impl IntoWarpReply for AddOrderError {
                      because we already stored different full app data for the same contract app \
                      data. This should be impossible."
                 );
-                shared::api::internal_error_reply()
+                crate::api::internal_error_reply()
             }
             AddOrderError::OrderNotFound(err) => err.into_warp_reply(),
             AddOrderError::InvalidAppData(err) => reply::with_status(
@@ -298,9 +298,9 @@ pub fn post_order(
 mod tests {
     use {
         super::*,
+        crate::api::response_body,
         model::order::{OrderCreation, OrderUid},
         serde_json::json,
-        shared::api::response_body,
         warp::{test::request, Reply},
     };
