+++ crates/shared/src/lib.rs
@@ -2,7 +2,6 @@
 pub mod macros;
 
 pub mod account_balances;
-pub mod api;
 pub mod arguments;
 pub mod bad_token;
 pub mod baseline_solver;
